touch `plone.api-install` to avoid running pip
When there is a separate target to only install `plone.api`, it runs pip every time to see if it is installed, which can take a second or two. It's an annoyance that adds up.

So I modified the Makefile to fake it, by making the target `plone.api-install` create a file with the same name. Make will skip that target's commands when a file with the name of the target exists, and, of course, it's not a phony target.

On the first run, `plone.api-install` will create the file `plone.api-install` and install plone.api. On subsequent runs, since a file with the name of the target exists, the target will be skipped.

If you want to purge the installation, use `make distclean`.

Note: while coming up with this, I was listening to the Peter Gabriel Live album, "I Have the Touch". Subliminal messages... 🤯