Remove the external link icon for links in submodules (#1916)
Remove the external link icon from Plone Sphinx Theme for links in submodules that refer to Plone 6 Documentation via Intersphinx.

Although these links have `external` in their class, they are actually internal to Plone 6 Documentation.

For example, see the index page of Volto UI.

https://6.docs.plone.org/volto/

And see that the external link icon for "Plone content management system" is correct, but the link for "First-time contributors" is actually internal but has the external link icon.